Add swiper to Vocabulary
