+++ src/components/app/main.tsx
@@ -1,0 +1,6 @@
+import ReactDOM from 'react-dom/client';
+import Provider from "../providers/providers.tsx";
+
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+    <Provider />,
+);
