+++ src/components/providers/providers.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import App from "../app/app.tsx";
+import {ChakraUIProvider} from "../../shared/ui/chakra/chakra-provider.ui.tsx";
+import {HashRouter} from "react-router-dom";
+
+
+const Provider: React.FC = () => {
+    return (
+        <ChakraUIProvider>
+            <HashRouter>
+                <App/>
+            </HashRouter>
+        </ChakraUIProvider>
+    );
+}
+
+export default Provider;
