+++ src/pages/game-page/answers/index.ts
