+++ src/pages/game-page/congratulation/index.ts
