+++ src/pages/routers/routers.tsx
@@ -3,19 +3,18 @@ import {
     AUTH_LINK,
     DICTIONARY_LINK,
     GAME_LINK,
-    HELP_LINK,
     HOME_LINK,
     NOT_FOUND_LINK
-} from "../../shared/constants-ui.ts";
+} from "../../shared/constants-link.ts";
 import {lazy, Suspense} from "react";
 import Spinner from "../../widgets/spinners/spinner.tsx";
 
 const GamePage = lazy(() => import('../game-page/game-page')) as any;
 const HomePage = lazy(() => import('../home-page/home-page')) as any;
 const AuthPage = lazy(() => import('../auth-page/auth-page')) as any;
-const DictionariesPage = lazy(() => import('../dictionaries-page/dictionaries-page')) as any;
+// const DictionariesPage = lazy(() => import('../dictionaries-page/dictionaries-page')) as any;
 const NotFoundPage = lazy(() => import('../not-found-page/not-found-page')) as any;
-const HelpPage = lazy(() => import('../help-page/help-page')) as any;
+const VocabulariesPage = lazy(() => import('../vocabularies-page/vocabularies-page')) as any;
 
 export const Routers = () => {
     const location = useLocation()
@@ -27,8 +26,7 @@ export const Routers = () => {
             <Route path={HOME_LINK} element={<Suspense fallback={<Spinner/>}><HomePage/></Suspense>}/>
             <Route path={GAME_LINK} element={<Suspense fallback={<Spinner/>}><GamePage/></Suspense>}/>
             <Route path={AUTH_LINK} element={<Suspense fallback={<Spinner/>}><AuthPage/></Suspense>}/>
-            <Route path={HELP_LINK} element={<Suspense fallback={<Spinner/>}><HelpPage/></Suspense>}/>
-            <Route path={DICTIONARY_LINK} element={<Suspense fallback={<Spinner/>}><DictionariesPage/></Suspense>}/>
+            <Route path={DICTIONARY_LINK} element={<Suspense fallback={<Spinner/>}><VocabulariesPage/></Suspense>}/>
         </Routes>
     )
 }
