+++ src/pages/vocabularies-page/dict-modal/index.ts
