+++ src/pages/vocabularies-page/list-of-vocabulary/index.ts
@@ -1,0 +1,2 @@
+export {ListOfVocabulary} from "./list-of-vocabulary.ui.tsx";
+export {Row} from "./row-of-list.tsx"
