+++ src/pages/vocabularies-page/vocabulries-swiper/vocabularies-swiper.css
@@ -1,0 +1,26 @@
+.swiper {
+    width: 100%;
+    height: 100%;
+    max-width: 95vw;
+
+}
+
+.swiper-slide {
+    text-align: center;
+    font-size: 50px;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding-bottom: 12%;
+    padding-top:2%;
+}
+
+
+.swiper-button-next {
+    right: 25px;
+}
+
+.swiper-button-prev {
+    left: 10px;
+}
