+++ src/shared/store/constants-store/easy-vocabulary.ts
@@ -1,0 +1,186 @@
+import {IVocabularyItem} from "../../types.ts";
+
+export const easyVocabulary:IVocabularyItem[]=[
+    {
+        id: "0",
+        word: "hello",
+        translate: "привет",
+        learning: 0,
+        popular: 100
+    },
+    {
+        id: "1",
+        word: "goodbye",
+        translate: "пока",
+        learning: 0,
+        popular: 90
+    },
+    {
+        id: "2",
+        word: "thank you",
+        translate: "спасибо",
+        learning: 0,
+        popular: 80
+    },
+    {
+        id: "3",
+        word: "good",
+        translate: "хорошо",
+        learning: 0,
+        popular: 70
+    },
+    {
+        id: "4",
+        word: "bad",
+        translate: "плохо",
+        learning: 0,
+        popular: 60
+    },
+    {
+        id: "5",
+        word: "yes",
+        translate: "да",
+        learning: 0,
+        popular: 50
+    },
+    {
+        id: "6",
+        word: "no",
+        translate: "нет",
+        learning: 0,
+        popular: 40
+    },
+    {
+        id: "7",
+        word: "maybe",
+        translate: "может быть",
+        learning: 0,
+        popular: 30
+    },
+    {
+        id: "8",
+        word: "please",
+        translate: "пожалуйста",
+        learning: 0,
+        popular: 20
+    },
+    {
+        id: "9",
+        word: "thank you very much",
+        translate: "спасибо большое",
+        learning: 0,
+        popular: 10
+    },
+    {
+        id: "10",
+        word: "in the end",
+        translate: "в конце концов",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "11",
+        word: "in the beginning",
+        translate: "в начале начал",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "12",
+        word: "in the middle",
+        translate: "в середине",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "13",
+        word: "in the evening",
+        translate: "в вечернее время",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "14",
+        word: "in the morning",
+        translate: "в утреннее время",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "15",
+        word: "in the afternoon",
+        translate: "в дней после обеда",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "16",
+        word: "in the night",
+        translate: "в ночное время",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "17",
+        word: "living room",
+        translate: "гостиная",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "18",
+        word: "kitchen",
+        translate: "кухня",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "19",
+        word: "bedroom",
+        translate: "спальня",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "20",
+        word: "bathroom",
+        translate: "ванная",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "21",
+        word: "toilet",
+        translate: "туалет",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "22",
+        word: "toilet paper",
+        translate: "бумажка",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "23",
+        word: "toilet brush",
+        translate: "ершик",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "24",
+        word: "toothpaste",
+        translate: "зубная паста",
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: "25",
+        word: "toothbrush",
+        translate: "зубная щетка",
+        learning: 0,
+        popular: 0
+    }
+]
