+++ src/shared/store/constants-store/lingvo-vocabulary.ts
@@ -1,0 +1,242 @@
+import {IVocabularyItem} from "../../types.ts";
+
+export const lingvoVocabulary: IVocabularyItem [] = [
+    {
+        id: "0",
+        word: "enough",
+        translate: "довольно, достаточно",
+        learning: 0,
+        popular: 60
+    },
+    {
+        id: "1",
+        word: "mention",
+        translate: "упоминать",
+        learning: 0,
+        popular: 70
+    },
+    {
+        id: "2",
+        word: "stuff",
+        translate: "вещи",
+        learning: 0,
+        popular: 50
+    },
+    {
+        id: "3",
+        word: "errand",
+        translate: "поручение",
+        learning: 0,
+        popular: 40
+    },
+    {
+        id: '4',
+        word: 'none of your business',
+        translate: 'не твое дело',
+        learning: 0,
+        popular: 30
+    },
+    {
+        id: '5',
+        word: 'particular',
+        translate: 'особый, конкретный',
+        learning: 0,
+        popular: 20
+    },
+    {
+        id: '6',
+        word: 'what\'s the metter',
+        translate: 'что случилось',
+        learning: 0,
+        popular: 10
+    },
+    {
+        id: '7',
+        word: 'likewise',
+        translate: 'взаимно, также',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '8',
+        word: 'rush around',
+        translate: 'носится по дому',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '9',
+        word: 'thought',
+        translate: 'мысль',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '10',
+        word: 'used to',
+        translate: 'привычный',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '11',
+        word: 'a bouquet',
+        translate: 'букет',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '12',
+        word: 'a branch of flowers',
+        translate: 'букет цветов',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '13',
+        word: 'deal with',
+        translate: 'иметь дело с',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '14',
+        word: 'whenever',
+        translate: 'каждый раз',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '15',
+        word: 'I used to',
+        translate: 'раньше я',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '16',
+        word: 'on my own',
+        translate: 'сам по себе',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '17',
+        word: 'I can\'t say',
+        translate: 'не могу сказать',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '18',
+        word: 'to sleep well',
+        translate: 'чтобы хорошо спать',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '19',
+        word: 'holly cow',
+        translate: 'святая корова...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '20',
+        word: 'holly moly',
+        translate: 'блин блинский...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '21',
+        word: 'Jesus christ',
+        translate: 'О божечки...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '22',
+        word: 'stay up later',
+        translate: 'засиживаться до поздна.',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '23',
+        word: 'to my shame',
+        translate: 'к моему стыду.',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '24',
+        word: 'set of',
+        translate: 'отправиться, начать что-то',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '25',
+        word: 'chorse',
+        translate: 'курс',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '26',
+        word: 'flatmates',
+        translate: 'тот кто живет с вами в дной квартире.',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '27',
+        word: 'feastive',
+        translate: 'праздничный',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '28',
+        word: 'One thing I don\'t like about is...',
+        translate: 'Единственное, что мне не нравится, это...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '29',
+        word: 'But on the other hand...',
+        translate: 'Но с другйо стороны...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '30',
+        word: 'However ...',
+        translate: 'однако...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '31',
+        word: 'At the same time ...',
+        translate: 'В то же время ...',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '32',
+        word: 'it is a big pleasure for me',
+        translate: 'это большая честь для меня.',
+        learning: 0,
+        popular: 0
+    },
+    {
+        id: '33',
+        word: 'its pleasure',
+        translate: 'с радостью',
+        learning: 0,
+        popular: 0
+    },
+]
