--- src/widgets/list-of-dictionary/index.ts
@@ -1,2 +1,0 @@
-export {ListOfDictionary} from "./list-of-dictionary.ui";
-export {Row} from "./row-of-list"
